+++ src/appsettings.h
@@ -222,11 +222,11 @@ public:
     { return "Ctrl+Shift+C"; }
 
     // Buttons
-    QOnlineTranslator::Language buttonLanguage(LangButtonGroup *group, int id);
-    void setButtonLanguage(LangButtonGroup *group, int id, QOnlineTranslator::Language lang);
+    QOnlineTranslator::Language buttonLanguage(LangButtonGroup &group, int id);
+    void setButtonLanguage(LangButtonGroup &group, int id, QOnlineTranslator::Language lang);
 
-    int checkedButton(LangButtonGroup *group);
-    void setCheckedButton(LangButtonGroup *group, int id);
+    int checkedButton(LangButtonGroup &group);
+    void setCheckedButton(LangButtonGroup &group, int id);
 
     // Main window settings
     QByteArray mainWindowGeometry();